deploy/az_f/bankmap/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
git+https://github.com/airenas/re-bankmap.git@v0.1.93-beta#egg=bankmap
